Remove obsolete Page and Utils classes
